sessions/session1/requirement.txt
pandas
psycopg2-binary
pytest
pytest-cov
python-dotenv
openpyxl
